backend/pyproject.toml
[tool.poetry]
name = "my django project"
version = "0.1.0"
description = ""
authors = ["노석현 <shtjrgus010@github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
black = "^24.10.0"
djangorestframework = "^3.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
